added CRUD for event

+++ src/event/event.service.ts
@@ -1,19 +1,30 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { UserDto } from '../user/dto/user.dto';
 import { Repository } from 'typeorm';
 import { CreateEventDto } from './dto/create-event.dto';
 import { UpdateEventDto } from './dto/update-event.dto';
 import { Event } from './entities/event.entity';
+import { EventDto } from './dto/event.dto';
+import { UserService } from '../user/user.service';
+import { toEventDto } from '../shared/mapper';
 
 @Injectable()
 export class EventService {
   constructor(
     @InjectRepository(Event)
     private readonly eventRepository: Repository<Event>,
+    private readonly userService: UserService,
   ) {}
 
-  create(createEventDto: CreateEventDto): Promise<Event> {
-    const event = new Event();
+  async create(
+    { username }: UserDto,
+    createEventDto: CreateEventDto,
+  ): Promise<EventDto> {
+    // get user from the db
+    const organizer = await this.userService.findOne({ where: { username } });
+
+    let event = new Event();
     event.title = createEventDto.title;
     event.capacity = createEventDto.capacity;
     event.isPrivate = createEventDto.isPrivate;
@@ -21,30 +32,91 @@ export class EventService {
     event.startDate = createEventDto.startDate;
     event.endDate = createEventDto.endDate;
 
-    return this.eventRepository.save(event);
+    event = this.eventRepository.create({ ...event, organizer });
+
+    await this.eventRepository.save(event);
+
+    return toEventDto(event);
   }
 
-  findAll() {
-    return this.eventRepository.find();
+  async findAll() {
+    const events = await this.eventRepository.find({
+      relations: ['organizer', 'members'],
+    });
+    return events.map((event) => toEventDto(event));
   }
 
   findOne(id: string) {
-    return this.eventRepository.findOneBy({ eventId: id });
+    return this.eventRepository.findOne({
+      where: { eventId: id },
+      relations: ['organizer', 'members'],
+    });
   }
 
-  update(id: string, updateEventDto: UpdateEventDto): Promise<Event> {
-    const event = new Event();
-    event.title = updateEventDto.title;
-    event.capacity = updateEventDto.capacity;
-    event.isPrivate = updateEventDto.isPrivate;
-    event.password = updateEventDto.password;
-    event.startDate = updateEventDto.startDate;
-    event.endDate = updateEventDto.endDate;
+  async update(
+    id: string,
+    { username }: UserDto,
+    updateEventDto: UpdateEventDto,
+  ): Promise<EventDto> {
+    const { title, capacity, isPrivate, password, startDate, endDate } =
+      updateEventDto;
+    let event: Event = await this.eventRepository.findOne({
+      where: { eventId: id },
+      relations: ['organizer'],
+    });
 
-    return this.eventRepository.save({ eventId: id, ...event });
+    if (!event) {
+      throw new HttpException(`Event doesn't exist`, HttpStatus.BAD_REQUEST);
+    }
+
+    // check permission
+    if (event.organizer.username !== username) {
+      throw new HttpException(
+        `Cannot update this event, you are not the organizer`,
+        HttpStatus.UNAUTHORIZED,
+      );
+    }
+
+    // update
+    await this.eventRepository.update(
+      { eventId: id },
+      {
+        title,
+        capacity,
+        isPrivate,
+        password,
+        startDate,
+        endDate,
+      },
+    );
+
+    // re-query
+    event = await this.eventRepository.findOne({
+      where: { eventId: id },
+      relations: ['organizer', 'members'],
+    });
+
+    return toEventDto(event);
   }
 
-  async remove(id: string): Promise<void> {
+  async remove(id: string, { username }: UserDto): Promise<void> {
+    const event: Event = await this.eventRepository.findOne({
+      where: { eventId: id },
+      relations: ['organizer'],
+    });
+
+    if (!event) {
+      throw new HttpException(`Event doesn't exist`, HttpStatus.BAD_REQUEST);
+    }
+
+    // check permission
+    if (event.organizer.username !== username) {
+      throw new HttpException(
+        `Cannot update this event, you are not the organizer`,
+        HttpStatus.UNAUTHORIZED,
+      );
+    }
+
     await this.eventRepository.delete({ eventId: id });
   }
 }
